+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Reported by @samwcoulter in https://github.com/razzmatazz/csharp-language-server/issues/289
 * Upgrade Roslyn libs to 5.0.0
 * csharp-ls will now use and require the .NET 10 runtime/SDK
+* Fix how title is formatted for nested code actions
+  - https://github.com/razzmatazz/csharp-language-server/pull/295
 
 **Full Changelog**: https://github.com/razzmatazz/csharp-language-server/compare/0.20.0...main